// src/app/core/models/property-features.interface.ts
export interface PropertyFeatures  {
  [key: string]: string;
  AC: string;
  Accessability: string;
  WindowBars: string;
  WaterHeating: string;
  Elevator: string;
  Roommates: string;
  Furniture: string;
  Unit: string;
  KosherKitchen: string;
  Pets: string;
  Renovated: string;
  Mamad: string;
  TadiranAC: string;
  Storage: string;
}